Slight problem with total price
The field for total cost does not disappear again when zero activities are selected. I can get it to remove when you uncheck 1 single box, but it removes even if some activities stay seleccted. 
As is, the total displays the correct amount, but displays zero when none are selected (as opposed to the total field disappearing).